requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas
numpy
scikit-learn
sentence-transformers
torch
firebase-admin
python-dotenv